refactor(ui): Extract calendar tooltip logic into useDelayedTooltip hook
