+++ airflow-core/src/airflow/ui/src/pages/Dag/Calendar/DailyCalendarView.tsx
@@ -37,12 +37,12 @@
  */
 import { Box, Text } from "@chakra-ui/react";
 import dayjs from "dayjs";
-import { useRef } from "react";
 
 import type { CalendarTimeRangeResponse } from "openapi/requests/types.gen";
 
 import { CalendarTooltip } from "./CalendarTooltip";
 import { createTooltipContent, generateDailyCalendarData, getCalendarCellColor } from "./calendarUtils";
+import { useDelayedTooltip } from "./useDelayedTooltip";
 
 type Props = {
   readonly cellSize: number;
@@ -52,39 +52,7 @@ type Props = {
 
 export const DailyCalendarView = ({ cellSize, data, selectedYear }: Props) => {
   const dailyData = generateDailyCalendarData(data, selectedYear);
-  const debounceTimeoutRef = useRef<NodeJS.Timeout | undefined>(undefined);
-  const activeTooltipRef = useRef<HTMLElement | undefined>(undefined);
-
-  const handleMouseEnter = (event: React.MouseEvent<HTMLDivElement>) => {
-    if (debounceTimeoutRef.current) {
-      clearTimeout(debounceTimeoutRef.current);
-    }
-
-    const tooltipElement = event.currentTarget.querySelector("[data-tooltip]");
-
-    if (tooltipElement) {
-      activeTooltipRef.current = tooltipElement as HTMLElement;
-      debounceTimeoutRef.current = setTimeout(() => {
-        if (activeTooltipRef.current) {
-          activeTooltipRef.current.style.opacity = "1";
-          activeTooltipRef.current.style.visibility = "visible";
-        }
-      }, 200);
-    }
-  };
-
-  const handleMouseLeave = () => {
-    if (debounceTimeoutRef.current) {
-      clearTimeout(debounceTimeoutRef.current);
-      debounceTimeoutRef.current = undefined;
-    }
-
-    if (activeTooltipRef.current) {
-      activeTooltipRef.current.style.opacity = "0";
-      activeTooltipRef.current.style.visibility = "hidden";
-      activeTooltipRef.current = undefined;
-    }
-  };
+  const { handleMouseEnter, handleMouseLeave } = useDelayedTooltip();
 
   return (
     <Box mb={4}>
