+++ airflow-core/src/airflow/ui/src/pages/Dag/Calendar/HourlyCalendarView.tsx
@@ -38,13 +38,13 @@
 import { Box, Text } from "@chakra-ui/react";
 import dayjs from "dayjs";
 import isSameOrBefore from "dayjs/plugin/isSameOrBefore";
-import { useRef } from "react";
 import { useTranslation } from "react-i18next";
 
 import type { CalendarTimeRangeResponse } from "openapi/requests/types.gen";
 
 import { CalendarTooltip } from "./CalendarTooltip";
 import { createTooltipContent, generateHourlyCalendarData, getCalendarCellColor } from "./calendarUtils";
+import { useDelayedTooltip } from "./useDelayedTooltip";
 
 dayjs.extend(isSameOrBefore);
 
@@ -58,39 +58,7 @@ type Props = {
 export const HourlyCalendarView = ({ cellSize, data, selectedMonth, selectedYear }: Props) => {
   const { t: translate } = useTranslation("dag");
   const hourlyData = generateHourlyCalendarData(data, selectedYear, selectedMonth);
-  const debounceTimeoutRef = useRef<NodeJS.Timeout | undefined>(undefined);
-  const activeTooltipRef = useRef<HTMLElement | undefined>(undefined);
-
-  const handleMouseEnter = (event: React.MouseEvent<HTMLDivElement>) => {
-    if (debounceTimeoutRef.current) {
-      clearTimeout(debounceTimeoutRef.current);
-    }
-
-    const tooltipElement = event.currentTarget.querySelector("[data-tooltip]");
-
-    if (tooltipElement) {
-      activeTooltipRef.current = tooltipElement as HTMLElement;
-      debounceTimeoutRef.current = setTimeout(() => {
-        if (activeTooltipRef.current) {
-          activeTooltipRef.current.style.opacity = "1";
-          activeTooltipRef.current.style.visibility = "visible";
-        }
-      }, 200);
-    }
-  };
-
-  const handleMouseLeave = () => {
-    if (debounceTimeoutRef.current) {
-      clearTimeout(debounceTimeoutRef.current);
-      debounceTimeoutRef.current = undefined;
-    }
-
-    if (activeTooltipRef.current) {
-      activeTooltipRef.current.style.opacity = "0";
-      activeTooltipRef.current.style.visibility = "hidden";
-      activeTooltipRef.current = undefined;
-    }
-  };
+  const { handleMouseEnter, handleMouseLeave } = useDelayedTooltip();
 
   return (
     <Box mb={4}>
